// src/types/movies.ts
export interface MoviesResponse {
    Title: string;
    Year: string;
    imdbID: string;
    Type: string;
    Poster: string;
}


interface RatesResponse {
    Source: string;
    Value: string;
}


export interface MoviesDetailsResponse {
    Title: string;
    Year: string;
    Rated: string;
    Released: string;
    Runtime: string;
    Genre: string;
    Director: string;
    Writer: string;
    Actors: string;
    Plot: string;
    Language: string;
    Country: string;
    Awards: string;
    Poster: string;
    Ratings: RatesResponse[];
    Metascore: string;
    imdbRating: string;
    imdbVotes: string;
    imdbID: string;
    Type: string;
    DVD: string;
    BoxOffice: string;
    Production: string;
    Website: string;
    Response: string;
}